requirements.txt
tensorflow~=2.2.0
numpy~=1.20.2
matplotlib~=3.4.1
Pillow~=8.1.2
PyQt5~=5.15.4
lxml~=4.6.3
labelImg~=1.8.4
setuptools~=54.2.0
pymusicxml==0.5.3
opencv-python==4.5.1.48
tqdm==4.60.0